[MOB-5337] - InAppDisplayInterval resolution
1. Replace instances of retryInterval to moveToForegroundSyncInterval.
2. Map inAppDisplayInterval to moveToForegroundSyncInterval rather than converting to retryInterval

Context:
The configurable inAppDisplayInteval maps to variable - retryInterval which is used to check if its ready to show inapp after first inapp was displayed. However, the same property is not used when app switches from background to foreground. During this time, a variable - moveToForegroundSyncInterval is used which is redundant and is never changed.