[MOB-9247] - Bug fix for no push on APNS SANDBOX
Logic explanation
SDK checks if the target is a simulator
if so > APNS SANDBOX
if not > read the file 'embedded.provision'
Check if certain value exist to mark it as APNSSANDBOX
If not, default it to PRODUCTION APNS.

Using Xcode 16 beta with iOS 18 beta, we discovered that file reading was failing with normal ASCII encoding.

Diff explanation -
File was able to type to data easily.
Once the data was available, now we try to read the file with all encoding available to us now. If none of them works, we print ITBLError.
The first one to decode the file proceeds with scan operation - finding the right tags and keywords.

- Tested with Xcode GA and Beta and iOS 17 and iOS 18 devices